+++ SystemWindow.lua
@@ -1,5 +1,5 @@
 local SystemWindow = {
-    VERSION = "1.1.0"
+    VERSION = "1.2.0"
 }
 
 function SystemWindow.configure(options)