+++ examples/minimal.lua
@@ -39,9 +39,15 @@ end
 function love.update(dt) windows:update(dt) end
 function love.draw() windows:draw() end
 function love.resize(w, h) windows:resize(w, h) end
-function love.mousepressed(...) windows:mousepressed(...) end
-function love.mousereleased(...) windows:mousereleased(...) end
-function love.mousemoved(...) windows:mousemoved(...) end
+function love.mousepressed(x, y, button, istouch, presses)
+    if not istouch then windows:mousepressed(x, y, button, false, presses) end
+end
+function love.mousereleased(x, y, button, istouch, presses)
+    if not istouch then windows:mousereleased(x, y, button, false, presses) end
+end
+function love.mousemoved(x, y, dx, dy, istouch)
+    if not istouch then windows:mousemoved(x, y, dx, dy, false) end
+end
 function love.wheelmoved(...) windows:wheelmoved(...) end
 function love.touchpressed(...) windows:touchpressed(...) end
 function love.touchmoved(...) windows:touchmoved(...) end